* origin/master: (108 commits)
  add non-bogus features to test dataset, include small test of such, change assert statements to those common in nose
  reduce convolution in GNBWeights computation, some renaming for better readability
  delete and/or improve comments
  move import to other imports
  revert previous introduction or deletion of whitespace
  extend gnb sensitivity test to test handling of zero variances and tagging/untagging
  handle cases of zero variance in features
  transform mapped numericals into ints for slicing
  change unit test to work with normal_feature_dataset, change raise assertion
  delete superflous datagenerator and its import
  GNB sensitivity analyzer implemented
  get AttributeMap import to use in GNB
  add some simple tests for current version of sensitivity analysis
  create tiny dataset with two labels & 2 features for simple unit test
  this commit just deletes or creates whitespace
  few GNB related imports in other files
  add GNBWeights class with first attempt of weight computation -- so far, clf.get_sensitivity_analyzer() only return a <GNBWeights> object instead of a dataset though
  add conditional get_sensitivity_analyser method to GNB class
  add tag if linear
  change necessary imports for GNB
  ...

 Conflicts:  largely just kept the master version
	mvpa2/base/externals.py
	mvpa2/base/info.py
	mvpa2/measures/searchlight.py

+++ mvpa2/tests/test_surfing_surface.py
@@ -138,7 +138,7 @@ class SurfingSurfaceTests(unittest.TestCase):
     @staticmethod
     def set_nan_to_surface_vertices(s, nan_ratio=.05):
         # make some vertices NaN (as might be the case for flat surfaces)
-        nan_count = np.ceil(s.nvertices * nan_ratio)
+        nan_count = int(np.ceil(s.nvertices * nan_ratio))
         nan_vertices_ids = np.random.random_integers(s.nvertices,
                                                      size=(nan_count,)) - 1
         vertices_noisy = s.vertices + 0.
